refactor: test 로직상의 mockWebServer를 설정하는 공통 부분 추출
