+++ src/test/java/kimsy/groceryapi/product/MockWebServerTest.java
@@ -1,0 +1,21 @@
+package kimsy.groceryapi.product;
+
+import java.io.IOException;
+import okhttp3.mockwebserver.MockWebServer;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+
+public abstract class MockWebServerTest {
+    protected static MockWebServer mockWebServer;
+
+    @BeforeAll
+    static void beforeAll() throws IOException {
+        mockWebServer = new MockWebServer();
+        mockWebServer.start();
+    }
+
+    @AfterAll
+    static void afterAll() throws IOException {
+        mockWebServer.shutdown();
+    }
+}
